+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - *specfk* method for SUdata()
 - *wiggle* method for SUdata()
 - Standard Genetic Algorithm implementation (mu, pc, pm)
+- Centroidal Voronoi tessellation option (McQueen algorithm) in *Swarm.init_particles*
+- Centroidal Voronoi tessellation option (McQueen algorithm) in *Genalg.init_chromosome*
 
 ### Modified
 - SU header structure now fits SU/CWP format (before SEG-Y rev 1)